fix: Resolve circular deps
