+++ packages/web-core/src/components/history.ts
@@ -1,16 +1,27 @@
 import { v4 } from 'uuid';
 
-import { LogsEventLogDetail } from '../../proto/ai/inworld/packets/packets.pb';
 import { DEFAULT_USER_NAME } from '../common/constants';
 import {
-  ConversationMapItem,
-  Extension,
+  ConversationState,
   InworlControlAction,
-  LogLevel,
-  TaskParameter,
-  TriggerParameter,
   User,
 } from '../common/data_structures';
+import {
+  ConvesationInterface,
+  Extension,
+} from '../common/data_structures/extension';
+import {
+  CHAT_HISTORY_TYPE,
+  HistoryInteractionEnd,
+  HistoryItem,
+  HistoryItemActor,
+  HistoryItemConversationUpdate,
+  HistoryItemLogEvent,
+  HistoryItemNarratedAction,
+  HistoryItemSceneChange,
+  HistoryItemTaskEvent,
+  HistoryItemTriggerEvent,
+} from '../common/data_structures/history';
 import { Character } from '../entities/character.entity';
 import { EmotionEvent } from '../entities/packets/emotion/emotion.entity';
 import { InworldPacket } from '../entities/packets/inworld_packet.entity';
@@ -26,114 +37,6 @@ interface InworldHistoryAddProps<InworldPacketT> {
   fromHistoryCharacter?: Character;
 }
 
-export enum CHAT_HISTORY_TYPE {
-  ACTOR = 'actor',
-  CONVERSATION_UPDATE = 'conversation_update',
-  INTERACTION_END = 'interaction_end',
-  LOG_EVENT = 'log_event',
-  NARRATED_ACTION = 'narrated_action',
-  SCENE_CHANGE = 'scene_change',
-  TRIGGER_EVENT = 'trigger_event',
-  TASK_EVENT = 'task_event',
-}
-
-export interface HistoryItemBase {
-  date: Date;
-  id: string;
-  scene: string;
-  interactionId?: string;
-  source: Actor;
-  type: CHAT_HISTORY_TYPE;
-  fromHistory?: boolean;
-  conversationId: string;
-}
-
-export interface HistoryItemActor extends HistoryItemBase {
-  type: CHAT_HISTORY_TYPE.ACTOR;
-  text?: string;
-  emotions?: EmotionEvent;
-  isRecognizing?: boolean;
-  // TODO: Remove this field in the next major release.
-  character?: Character;
-  characters?: Character[];
-  correlationId?: string;
-}
-
-export interface HistoryItemTriggerEvent extends HistoryItemBase {
-  type: CHAT_HISTORY_TYPE.TRIGGER_EVENT;
-  name: string;
-  parameters: TriggerParameter[];
-  outgoing?: boolean;
-  correlationId?: string;
-}
-
-export interface HistoryItemTaskEvent extends HistoryItemBase {
-  type: CHAT_HISTORY_TYPE.TASK_EVENT;
-  name: string;
-  parameters: TaskParameter[];
-  outgoing?: boolean;
-  correlationId?: string;
-}
-
-export interface HistoryInteractionEnd extends HistoryItemBase {
-  type: CHAT_HISTORY_TYPE.INTERACTION_END;
-}
-
-export interface HistoryItemNarratedAction extends HistoryItemBase {
-  type: CHAT_HISTORY_TYPE.NARRATED_ACTION;
-  text?: string;
-  // TODO: Remove this field in the next major release.
-  character?: Character;
-  characters?: Character[];
-}
-
-export interface HistoryItemLogEvent extends HistoryItemBase {
-  type: CHAT_HISTORY_TYPE.LOG_EVENT;
-  text: string;
-  level: LogLevel;
-  metadata?: Record<string, string>;
-  details?: LogsEventLogDetail[];
-}
-
-export interface HistoryItemSceneChange {
-  date: Date;
-  id: string;
-  interactionId?: string;
-  source: Actor;
-  type: CHAT_HISTORY_TYPE.SCENE_CHANGE;
-  to?: string;
-  description?: string;
-  displayName?: string;
-  loadedCharacters?: Character[];
-  addedCharacters?: Character[];
-  removedCharacters?: Character[];
-  conversationId?: string;
-  fromHistory?: boolean;
-}
-
-export interface HistoryItemConversationUpdate {
-  date: Date;
-  id: string;
-  interactionId?: string;
-  source: Actor;
-  type: CHAT_HISTORY_TYPE.CONVERSATION_UPDATE;
-  conversationId?: string;
-  currentCharacters?: Character[];
-  addedCharacters?: Character[];
-  removedCharacters?: Character[];
-  fromHistory?: boolean;
-}
-
-export type HistoryItem =
-  | HistoryItemActor
-  | HistoryItemLogEvent
-  | HistoryItemTriggerEvent
-  | HistoryItemTaskEvent
-  | HistoryInteractionEnd
-  | HistoryItemNarratedAction
-  | HistoryItemSceneChange
-  | HistoryItemConversationUpdate;
-
 interface EmotionsMap {
   [key: string]: EmotionEvent;
 }
@@ -146,7 +49,13 @@ interface InworldHistoryProps<
   extension?: Extension<InworldPacketT, HistoryItemT>;
   user?: User;
   scene: string;
-  conversations: Map<string, ConversationMapItem<InworldPacketT>>;
+  conversations: Map<
+    string,
+    {
+      service: ConvesationInterface<InworldPacketT>;
+      state: ConversationState;
+    }
+  >;
 }
 
 interface ConversationItem {
@@ -167,7 +76,13 @@ export class InworldHistory<
   private emotions: EmotionsMap = {};
   private extension: Extension<InworldPacketT, HistoryItemT> | undefined;
   private conversationItems: ConversationItem[] = [];
-  private conversations: Map<string, ConversationMapItem<InworldPacketT>>;
+  private conversations: Map<
+    string,
+    {
+      service: ConvesationInterface<InworldPacketT>;
+      state: ConversationState;
+    }
+  >;
 
   constructor(props: InworldHistoryProps<InworldPacketT, HistoryItemT>) {
     if (props.extension) {
